requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf_yasg
requests